proj_docker/get_schedule_data/requirements.txt
kafka-python
requests
PyYAML
python-dotenv